Conflicts:
	tests/chainer_tests/links_tests/connection_tests/test_gru.py

+++ chainer/cuda.py
@@ -27,6 +27,8 @@ import os
 import warnings
 
 import numpy
+import six
+
 
 available = False
 cudnn_enabled = False
@@ -52,10 +54,11 @@ except Exception as e:
         pass  # for type testing
 
 if available:
+    _cudnn_disabled_by_user = int(os.environ.get('CHAINER_CUDNN', '1')) == 0
     try:
         import cupy.cudnn
         cudnn = cupy.cudnn
-        cudnn_enabled = int(os.environ.get('CHAINER_CUDNN', '1')) != 0
+        cudnn_enabled = not _cudnn_disabled_by_user
     except Exception as e:
         _resolution_error = e
 
@@ -80,11 +83,14 @@ def check_cuda_available():
                '(see https://github.com/pfnet/chainer#installation).')
         msg += str(_resolution_error)
         raise RuntimeError(msg)
-    if not cudnn_enabled:
+    if (not cudnn_enabled and
+            not _cudnn_disabled_by_user and
+            not getattr(check_cuda_available, '_already_warned', False)):
         warnings.warn(
             'cuDNN is not enabled.\n'
             'Please reinstall chainer after you install cudnn\n'
             '(see https://github.com/pfnet/chainer#installation).')
+        check_cuda_available._already_warned = True
 
 
 class DummyDeviceType(object):
@@ -94,6 +100,9 @@ class DummyDeviceType(object):
     This class is used to represent CPU device.
 
     """
+
+    id = -1
+
     def __int__(self):
         return -1
 
@@ -139,13 +148,12 @@ def get_device(*args):
     protocol of Python for the *with* statement.
 
     Args:
-        args: Values to specify a GPU device. :class:`numpy.ndarray` objects
-            are skipped. If all arguments are :class:`numpy.ndarray` objects,
-            it returns a dummy device object. Otherwise, the first
-            non-:mod:`numpy` object is used to select a device. If it is a
-            :class:`cupy.ndarray` object, its device is returned. Otherwise,
-            the argument is passed to the initializer of
-            :class:`~cupy.cuda.Device` and it is returned.
+        args: Values to specify a GPU device. The first integer or
+            :class:`cupy.ndarray` object is used to select a device. If it is
+            an integer, the corresponding device is returned. If it is a CuPy
+            array, the device on which this array reside is returned. If any
+            arguments are neither integers nor CuPy arrays, a dummy device
+            object representing CPU is returned.
 
     Returns:
         Device object specified by given ``args``.
@@ -155,16 +163,13 @@ def get_device(*args):
 
     """
     for arg in args:
-        if arg is None:
-            continue
-        if not isinstance(arg, numpy.ndarray):
+        if type(arg) in six.integer_types:
             check_cuda_available()
-            if isinstance(arg, cupy.ndarray):
-                if arg.device is None:
-                    continue
-                return arg.device
-            else:
-                return Device(arg)
+            return Device(arg)
+        if isinstance(arg, ndarray):
+            if arg.device is None:
+                continue
+            return arg.device
 
     return DummyDevice
 
@@ -179,7 +184,8 @@ def to_gpu(array, device=None, stream=None):
     Args:
         array: Array to be sent to GPU.
         device: Device specifier.
-        stream (cupy.cuda.Stream): CUDA stream.
+        stream (cupy.cuda.Stream): CUDA stream. If not ``None``, the copy runs
+            asynchronously.
 
     Returns:
         cupy.ndarray: Array on GPU.
@@ -190,14 +196,32 @@ def to_gpu(array, device=None, stream=None):
 
     """
     check_cuda_available()
-    assert stream is None  # TODO(beam2d): FIX IT
     with get_device(device):
-        dev_id = int(get_device(array))
-        if dev_id != -1 and dev_id != cupy.cuda.device.get_device_id():
-            # Need to make a copy when an array is copied to another device
-            return cupy.array(array, copy=True)
-        else:
+        array_dev = get_device(array)
+        if array_dev.id == cupy.cuda.device.get_device_id():
+            return array
+
+        if stream is not None:
+            ret = cupy.empty_like(array)
+            if array_dev.id == -1:
+                # cpu to gpu
+                src = array.copy(order='C')
+                ret.set(src, stream)
+            else:
+                # gpu to gpu
+                with array_dev:
+                    src = array.copy()
+                ret.data.copy_from_device_async(src.data, src.nbytes, stream)
+
+            # to hold a reference until the end of the asynchronous memcpy
+            stream.add_callback(lambda *x: None, (src, ret))
+            return ret
+
+        if array_dev.id == -1:
             return cupy.asarray(array)
+
+        # Need to make a copy when an array is copied to another device
+        return cupy.array(array, copy=True)
 
 
 def to_cpu(array, stream=None):
@@ -439,6 +463,18 @@ def memoize(for_each_device=False):
     return dummy_decorator
 
 
+def clear_memo():
+    """Clears the memoized results for all functions decorated by memoize.
+
+    This function works like :func:`cupy.clear_memo` as a counterpart for
+    :func:`chainer.cuda.memoize`. It can be used even if CUDA is not available.
+    In such a case, this function does nothing.
+
+    """
+    if available:
+        cupy.clear_memo()
+
+
 # ------------------------------------------------------------------------------
 # Kernel definition utility
 # ------------------------------------------------------------------------------
@@ -506,11 +542,26 @@ def get_array_module(*args):
 _max_workspace_size = 8 * 1024 * 1024
 
 
-# TODO(okuta): Write document
 def get_max_workspace_size():
+    """Gets the workspace size for cuDNN.
+
+    Check "cuDNN Library User Guide" for detail.
+
+    Returns:
+        int: The workspace size for cuDNN.
+
+    """
     return _max_workspace_size
 
 
 def set_max_workspace_size(size):
+    """Sets the workspace size for cuDNN.
+
+    Check "cuDNN Library User Guide" for detail.
+
+    Args:
+        size: The workspace size for cuDNN.
+
+    """
     global _max_workspace_size
     _max_workspace_size = size
